Fill perfdata to be able to print tree
