+++ thicket/ensemble.py
@@ -4,10 +4,8 @@
 # SPDX-License-Identifier: MIT
 
 from collections import OrderedDict
-import warnings
 
 from hatchet import GraphFrame
-import numpy as np
 import pandas as pd
 
 import thicket.helpers as helpers
@@ -16,6 +14,7 @@ from .utils import (
     validate_dataframe,
     verify_sorted_profile,
     verify_thicket_structures,
+    _fill_perfdata,
 )
 
 
@@ -187,9 +186,9 @@ class Ensemble:
             combined_th.profile = [new_mappings[prf] for prf in combined_th.profile]
             profile_mapping_cp = combined_th.profile_mapping.copy()
             for k, v in profile_mapping_cp.items():
-                combined_th.profile_mapping[
-                    new_mappings[k]
-                ] = combined_th.profile_mapping.pop(k)
+                combined_th.profile_mapping[new_mappings[k]] = (
+                    combined_th.profile_mapping.pop(k)
+                )
             combined_th.performance_cols = helpers._get_perf_columns(
                 combined_th.dataframe
             )
@@ -352,38 +351,6 @@ class Ensemble:
             unify_profile (list): profiles,
             unify_profile_mapping (dict): profile mapping
         """
-
-        def _fill_perfdata(df, numerical_fill_value=np.nan):
-            """Create full index for DataFrame and fill created rows with NaN's or None's where applicable.
-
-            Arguments:
-                df (DataFrame): DataFrame to fill missing rows in
-                numerical_fill_value (any): value to fill numerical rows with
-
-            Returns:
-                (DataFrame): filled DataFrame
-            """
-            try:
-                # Fill missing rows in dataframe with NaN's
-                df = df.reindex(
-                    pd.MultiIndex.from_product(df.index.levels),
-                    fill_value=numerical_fill_value,
-                )
-                # Replace "NaN" with "None" in columns of string type
-                for col in df.columns:
-                    if pd.api.types.is_string_dtype(df[col].dtype):
-                        df[col] = df[col].replace({numerical_fill_value: None})
-            except ValueError as e:
-                estr = str(e)
-                if estr == "cannot handle a non-unique multi-index!":
-                    warnings.warn(
-                        "Non-unique multi-index for DataFrame in _fill_perfdata. Cannot Fill missing rows.",
-                        RuntimeWarning,
-                    )
-                else:
-                    raise
-
-            return df
 
         # Add missing indicies to thickets
         helpers._resolve_missing_indicies(thickets)
